+++ cli/cmd/create.go
@@ -44,7 +44,8 @@ func NewCreateCmd() *cobra.Command {
 Built-in templates:
 	cartridge: a simple Cartridge application.
 	single_instance: Tarantool 3 application with a single instance configuration.
-	vshard_cluster: Tarantool 3 vshard cluster application.`,
+	vshard_cluster: Tarantool 3 vshard cluster application.
+	config_storage: Tarantool 3 cluster configuration storage.`,
 		Example: `
 # Create an application app1 from a template.
 
@@ -62,7 +63,11 @@ Built-in templates:
 
 # Create Tarantool 3 vshard cluster.
 
-    $ tt create vshard_cluster --name cluster_app`,
+    $ tt create vshard_cluster --name cluster_app
+
+# Create Tarantool 3 cluster configuration storage.
+
+    $ tt create config_storage --name cs_app`,
 	}
 
 	createCmd.Flags().StringVarP(&appName, "name", "n", "", "Application name")